+++ lib/bin/lwm2m_carrier/os/lwm2m_os.c
@@ -10,7 +10,7 @@
 #include <stdbool.h>
 #include <zephyr.h>
 #include <string.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 #include <at_notif.h>
 #include <at_cmd_parser/at_cmd_parser.h>
 #include <at_cmd_parser/at_params.h>
