+++ lib/lte_link_control/lte_lc.c
@@ -12,7 +12,7 @@
 #include <stdio.h>
 #include <device.h>
 #include <lte_lc.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 #include <at_cmd_parser/at_cmd_parser.h>
 #include <at_cmd_parser/at_params.h>
 #include <at_notif.h>
