+++ lib/modem_key_mgmt/modem_key_mgmt.c
@@ -7,7 +7,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <zephyr.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 #include <at_cmd_parser/at_cmd_parser.h>
 #include <at_cmd_parser/at_params.h>
 #include <bsd_limits.h>
