+++ samples/nrf9160/gps/src/main.c
@@ -8,7 +8,7 @@
 #include <nrf_socket.h>
 #include <net/socket.h>
 #include <stdio.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 #include <at_notif.h>
 
 #ifdef CONFIG_SUPL_CLIENT_LIB
