+++ samples/nrf9160/http_application_update/src/main.c
@@ -8,7 +8,7 @@
 #include <drivers/flash.h>
 #include <bsd.h>
 #include <lte_lc.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 #include <at_notif.h>
 #include <net/bsdlib.h>
 #include <net/fota_download.h>
