+++ samples/nrf9160/serial_lte_modem/src/slm_at_icmp.h
@@ -14,7 +14,7 @@
  */
 
 #include <zephyr/types.h>
-#include <at_cmd.h>
+#include <modem/at_cmd.h>
 
 /**
  * @brief ICMP AT command parser.
